+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 * WatchMonitor增加createAll方法
 * 增加NetUtil.getLocalhost
 * 优化ClassPathResource错误提示
+* 增加 ArrayUtil.toArray方法
 
 ### Bug修复
 * DateUtil中offsite修正为offset